Use int instead for comparing values

+++ src/main/java/org/folio/dew/batch/bursarfeesfines/service/BursarFilterAggregateEvaluator.java
@@ -39,7 +39,7 @@ public class BursarFilterAggregateEvaluator {
       }
       return compareHelper(
         filter.getCondition(),
-        totalAmount.floatValue(),
+        totalAmount.intValue(),
         filter.getAmount()
       );
     } else {
@@ -51,8 +51,8 @@ public class BursarFilterAggregateEvaluator {
   // TODO: determine if float is the desired type here?
   private final boolean compareHelper(
     ConditionEnum condition,
-    float amount,
-    float conditionAmount
+    int amount,
+    int conditionAmount
   ) {
     switch (condition) {
       case LESS_THAN:
